+++ TestProjects/HDRP_Tests/Assets/GraphicTests/Common/Scripts/LookAtCamera.cs
@@ -4,14 +4,14 @@ using UnityEngine;
 
 public class LookAtCamera : MonoBehaviour
 {
-	public new Camera camera;
+	public Camera targetCamera;
 	public Vector3 additionalRotation = new Vector3(0f, 180f, 0f);
 
 	[ContextMenu("Look At")]
 	public void LookAt()
 	{
-		if (camera == null && Camera.main == null) return;
+		if (targetCamera == null && Camera.main == null) return;
 
-		transform.rotation = Quaternion.LookRotation( ( (camera==null) ? Camera.main.transform.position : camera.transform.position ) - transform.position ) * Quaternion.Euler(additionalRotation);
+		transform.rotation = Quaternion.LookRotation( ( (targetCamera==null) ? Camera.main.transform.position : targetCamera.transform.position ) - transform.position ) * Quaternion.Euler(additionalRotation);
 	}
 }