+++ com.unity.render-pipelines.high-definition/Editor/Lighting/Reflection/Volume/InfluenceVolumeUI.Drawers.cs
@@ -8,23 +8,18 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
 
     partial class InfluenceVolumeUI
     {
-        //[TODO: planar / non planar will be redone in next PR]
-        internal static readonly CED.IDrawer SectionFoldoutShapePlanar;
-        internal static readonly CED.IDrawer SectionFoldoutShape;
         static readonly CED.IDrawer SectionShapeBoxPlanar = CED.Action((s, p, o) => Drawer_SectionShapeBox( s, p, o, false, false, false));
         static readonly CED.IDrawer SectionShapeBox = CED.Action((s, p, o) => Drawer_SectionShapeBox(s, p, o, true, true, true));
         static readonly CED.IDrawer SectionShapeSpherePlanar = CED.Action((s, p, o) => Drawer_SectionShapeSphere(s, p, o, false, false));
         static readonly CED.IDrawer SectionShapeSphere = CED.Action((s, p, o) => Drawer_SectionShapeSphere(s, p, o, true, true));
 
-        static InfluenceVolumeUI()
+        internal static CED.IDrawer InnerInspector(ReflectionProbeType type)
         {
-            SectionFoldoutShapePlanar = CED.Group(
-                    CED.FoldoutGroup(
-                        influenceVolumeHeader,
-                        (s, d, o) => s.isSectionExpandedShape,
-                        FoldoutOption.Indent,
+            switch (type)
+            {
+                case ReflectionProbeType.PlanarReflection:
+                    return CED.Group(
                         CED.Action(Drawer_InfluenceAdvancedSwitch),
-                        CED.space,
                         CED.Action(Drawer_FieldShapeType),
                         CED.FadeGroup(
                             (s, d, o, i) => s.IsSectionExpanded_Shape((InfluenceShape)i),
@@ -32,15 +27,10 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                             SectionShapeBoxPlanar,
                             SectionShapeSpherePlanar
                             )
-                        )
-                    );
-            SectionFoldoutShape = CED.Group(
-                    CED.FoldoutGroup(
-                        influenceVolumeHeader,
-                        (s, d, o) => s.isSectionExpandedShape,
-                        FoldoutOption.Indent,
+                        );
+                case ReflectionProbeType.ReflectionProbe:
+                    return CED.Group(
                         CED.Action(Drawer_InfluenceAdvancedSwitch),
-                        CED.space,
                         CED.Action(Drawer_FieldShapeType),
                         CED.FadeGroup(
                             (s, d, o, i) => s.IsSectionExpanded_Shape((InfluenceShape)i),
@@ -48,8 +38,10 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                             SectionShapeBox,
                             SectionShapeSphere
                             )
-                        )
-                    );
+                        );
+                default:
+                    throw new System.ArgumentException("Unknown probe type");
+            }
         }
 
         static void Drawer_FieldShapeType(InfluenceVolumeUI s, SerializedInfluenceVolume d, Editor o)
@@ -69,8 +61,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
                 }
 
                 bool advanced = d.editorAdvancedModeEnabled.boolValue;
-                advanced = !GUILayout.Toggle(!advanced, normalModeContent, EditorStyles.miniButtonLeft, GUILayout.Width(60f), GUILayout.ExpandWidth(false));
-                advanced = GUILayout.Toggle(advanced, advancedModeContent, EditorStyles.miniButtonRight, GUILayout.Width(60f), GUILayout.ExpandWidth(false));
+                advanced = GUILayout.Toggle(advanced, advancedModeContent, EditorStyles.miniButton, GUILayout.Width(60f), GUILayout.ExpandWidth(false));
                 if (d.editorAdvancedModeEnabled.boolValue ^ advanced)
                 {
                     d.editorAdvancedModeEnabled.boolValue = advanced;
@@ -99,7 +90,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             }
         }
 
-        static void Drawer_SectionShapeBox(InfluenceVolumeUI s, SerializedInfluenceVolume d, Editor o, bool drawOffset, bool drawNormal, bool drawFace)
+        static void Drawer_SectionShapeBox(InfluenceVolumeUI s, SerializedInfluenceVolume d, Editor o, bool drawAllOffset, bool drawNormal, bool drawFace)
         {
             bool advanced = d.editorAdvancedModeEnabled.boolValue;
             var maxFadeDistance = d.boxSize.vector3Value * 0.5f;
@@ -150,10 +141,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             HDProbeUI.Drawer_ToolBarButton(HDProbeUI.ToolBar.InfluenceShape, o, GUILayout.Width(28f), GUILayout.MinHeight(22f));
             EditorGUILayout.EndHorizontal();
 
-            if (drawOffset)
-            {
-                Drawer_Offset(s, d, o);
-            }
+            Drawer_Offset(s, d, o, drawAllOffset);
             
             GUILayout.Space(EditorGUIUtility.standardVerticalSpacing);
             
@@ -241,18 +229,15 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             GUILayout.EndVertical();
         }
 
-        static void Drawer_SectionShapeSphere(InfluenceVolumeUI s, SerializedInfluenceVolume d, Editor o, bool drawOffset, bool drawNormal)
+        static void Drawer_SectionShapeSphere(InfluenceVolumeUI s, SerializedInfluenceVolume d, Editor o, bool drawOffset, bool drawAllOffset)
         {
 
             EditorGUILayout.BeginHorizontal();
             EditorGUILayout.PropertyField(d.sphereRadius, radiusContent);
             HDProbeUI.Drawer_ToolBarButton(HDProbeUI.ToolBar.InfluenceShape, o, GUILayout.Width(28f), GUILayout.MinHeight(22f));
             EditorGUILayout.EndHorizontal();
-
-            if(drawOffset)
-            {
-                Drawer_Offset(s, d, o);
-            }
+            
+            Drawer_Offset(s, d, o, drawAllOffset);
 
             EditorGUILayout.Space();
             var maxBlendDistance = d.sphereRadius.floatValue;
@@ -267,7 +252,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             HDProbeUI.Drawer_ToolBarButton(HDProbeUI.ToolBar.Blend, o, GUILayout.ExpandHeight(true), GUILayout.Width(28f), GUILayout.MinHeight(22f), GUILayout.MaxHeight(EditorGUIUtility.singleLineHeight + 3));
             EditorGUILayout.EndHorizontal();
 
-            if (drawNormal)
+            if (drawAllOffset)
             {
                 EditorGUILayout.BeginHorizontal();
                 EditorGUI.BeginChangeCheck();
@@ -281,17 +266,32 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
             }
         }
 
-        static void Drawer_Offset(InfluenceVolumeUI s, SerializedInfluenceVolume d, Editor o)
+        static void Drawer_Offset(InfluenceVolumeUI s, SerializedInfluenceVolume d, Editor o, bool drawAllOffset)
         {
             EditorGUILayout.BeginHorizontal();
             EditorGUI.BeginChangeCheck();
-            EditorGUILayout.PropertyField(d.offset, offsetContent);
+            float y = 0f;
+            if (drawAllOffset)
+            {
+                EditorGUILayout.PropertyField(d.offset, offsetContent);
+            }
+            else
+            {
+                y = EditorGUILayout.FloatField(yOffsetContent, d.offset.vector3Value.y);
+            }
             if(EditorGUI.EndChangeCheck())
             {
                 //call the offset setter as it will update legacy reflection probe
                 HDProbeEditor editor = (HDProbeEditor)o;
                 InfluenceVolume influenceVolume = editor.GetTarget(editor.target).influenceVolume;
-                influenceVolume.offset = d.offset.vector3Value;
+                if (drawAllOffset)
+                {
+                    influenceVolume.offset = d.offset.vector3Value;
+                }
+                else
+                {
+                    influenceVolume.offset = new Vector3(0, y, 0);
+                }
             }
             HDProbeUI.Drawer_ToolBarButton(HDProbeUI.ToolBar.CapturePosition, o, GUILayout.Width(28f), GUILayout.MinHeight(22f));
             EditorGUILayout.EndHorizontal();