+++ com.unity.render-pipelines.high-definition/Editor/Material/Lit/ShaderGraph/HDLitSettingsView.cs
@@ -46,7 +46,7 @@ namespace UnityEditor.ShaderGraph.Drawing
                 ++indentLevel;
                 if (!m_Node.HasRefraction())
                 {
-                    ps.Add(new PropertyRow(CreateLabel("Blend Mode", indentLevel)), (row) =>
+                    ps.Add(new PropertyRow(CreateLabel("Blending Mode", indentLevel)), (row) =>
                     {
                         row.Add(new EnumField(HDLitMasterNode.AlphaModeLit.Additive), (field) =>
                         {
@@ -87,7 +87,7 @@ namespace UnityEditor.ShaderGraph.Drawing
                 {
                     ps.Add(new PropertyRow(CreateLabel("Refraction Model", indentLevel)), (row) =>
                     {
-                        row.Add(new EnumField(ScreenSpaceLighting.RefractionModel.None), (field) =>
+                        row.Add(new EnumField(ScreenSpaceRefraction.RefractionModel.None), (field) =>
                         {
                             field.value = m_Node.refractionModel;
                             field.OnValueChanged(ChangeRefractionModel);
@@ -339,7 +339,7 @@ namespace UnityEditor.ShaderGraph.Drawing
                 return;
 
             m_Node.owner.owner.RegisterCompleteObjectUndo("Refraction Model Change");
-            m_Node.refractionModel = (ScreenSpaceLighting.RefractionModel)evt.newValue;
+            m_Node.refractionModel = (ScreenSpaceRefraction.RefractionModel)evt.newValue;
         }
 
         void ChangeDistortion(ChangeEvent<bool> evt)