+++ com.unity.render-pipelines.high-definition/Editor/RenderPipeline/Settings/FrameSettingsUI.Skin.cs
@@ -10,6 +10,7 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
         const string renderingSettingsHeaderContent = "Rendering Settings";
         const string xrSettingsHeaderContent = "XR Settings";
         const string lightSettingsHeaderContent = "Lighting Settings";
+        const string asyncComputeSettingsHeaderContent = "Async Compute Settings";
         
         static readonly GUIContent transparentPrepassContent = CoreEditorUtils.GetContent("Transparent Prepass");
         static readonly GUIContent transparentPostpassContent = CoreEditorUtils.GetContent("Transparent Postpass");
@@ -21,7 +22,6 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
         static readonly GUIContent postprocessContent = CoreEditorUtils.GetContent("Postprocess");
         static readonly GUIContent litShaderModeContent = CoreEditorUtils.GetContent("Lit Shader Mode");
         static readonly GUIContent depthPrepassWithDeferredRenderingContent = CoreEditorUtils.GetContent("Depth Prepass With Deferred Rendering");
-        static readonly GUIContent asyncComputeContent = CoreEditorUtils.GetContent("Async Compute");
         static readonly GUIContent opaqueObjectsContent = CoreEditorUtils.GetContent("Opaque Objects");
         static readonly GUIContent transparentObjectsContent = CoreEditorUtils.GetContent("Transparent Objects");
         static readonly GUIContent realtimePlanarReflectionContent = CoreEditorUtils.GetContent("Enable Realtime Planar Reflection"); 
@@ -37,6 +37,12 @@ namespace UnityEditor.Experimental.Rendering.HDPipeline
         static readonly GUIContent volumetricContent = CoreEditorUtils.GetContent("Volumetric");
         static readonly GUIContent reprojectionForVolumetricsContent = CoreEditorUtils.GetContent("Reprojection For Volumetrics");
         static readonly GUIContent lightLayerContent = CoreEditorUtils.GetContent("LightLayers");
+
+        // Async compute
+        static readonly GUIContent asyncComputeContent = CoreEditorUtils.GetContent("Async Compute|This will have an effect only if target platform supports async compute.");
+        static readonly GUIContent lightListAsyncContent = CoreEditorUtils.GetContent("Build Light List in Async");
+        static readonly GUIContent SSRAsyncContent = CoreEditorUtils.GetContent("SSR in Async");
+        static readonly GUIContent SSAOAsyncContent = CoreEditorUtils.GetContent("SSAO in Async");
 
         static readonly GUIContent frameSettingsHeaderContent = CoreEditorUtils.GetContent("Frame Settings Override|Default FrameSettings are defined in HDRenderPipelineAsset.");
     }