+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/ScreenSpaceLighting/ScreenSpaceReflection.cs
@@ -4,8 +4,18 @@ using UnityEngine.Rendering;
 namespace UnityEngine.Experimental.Rendering.HDPipeline
 {
     [Serializable]
-    public class ScreenSpaceReflection : ScreenSpaceLighting
+    public class ScreenSpaceReflection : VolumeComponent
     {
+
+        public ClampedFloatParameter depthBufferThickness = new ClampedFloatParameter(0.01f, 0, 1);
+        public ClampedFloatParameter screenFadeDistance = new ClampedFloatParameter(0.1f, 0.0f, 1.0f);
+        public ClampedFloatParameter minSmoothness = new ClampedFloatParameter(0.9f, 0.0f, 1.0f);
+        public ClampedFloatParameter smoothnessFadeStart = new ClampedFloatParameter(0.9f, 0.0f, 1.0f);
+        public BoolParameter         reflectSky          = new BoolParameter(true);
+
+        public IntParameter rayMaxIterations = new IntParameter(32);
+
+
         static ScreenSpaceReflection s_Default = null;
         public static ScreenSpaceReflection @default
         {
@@ -20,31 +30,5 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             }
         }
 
-        public ClampedFloatParameter minSmoothness       = new ClampedFloatParameter(0.9f, 0.0f, 1.0f);
-        public ClampedFloatParameter smoothnessFadeStart = new ClampedFloatParameter(0.9f, 0.0f, 1.0f);
-        public BoolParameter         reflectSky          = new BoolParameter(true);
-
-        protected override void FetchIDs(
-            out int rayLevelID,
-            out int rayMinLevelID,
-            out int rayMaxLevelID,
-            out int rayMaxIterationsID,
-            out int depthBufferThicknessID,
-            out int screenWeightDistanceID,
-            out int rayMaxScreenDistanceID,
-            out int rayBlendScreenDistanceID,
-            out int rayMarchBehindObjectsID
-            )
-        {
-            rayLevelID = HDShaderIDs._SSReflectionRayLevel;
-            rayMinLevelID = HDShaderIDs._SSReflectionRayMinLevel;
-            rayMaxLevelID = HDShaderIDs._SSReflectionRayMaxLevel;
-            rayMaxIterationsID = HDShaderIDs._SSReflectionRayMaxIterations;
-            depthBufferThicknessID = HDShaderIDs._SSReflectionDepthBufferThickness;
-            screenWeightDistanceID = HDShaderIDs._SSReflectionInvScreenWeightDistance;
-            rayMaxScreenDistanceID = HDShaderIDs._SSReflectionRayMaxScreenDistance;
-            rayBlendScreenDistanceID = HDShaderIDs._SSReflectionRayBlendScreenDistance;
-            rayMarchBehindObjectsID = HDShaderIDs._SSReflectionRayMarchBehindObjects;
-        }
     }
 }